requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1
celery[redis]==5.3.6
redis==5.0.1
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
